Project splitted into modules
